Rename project and packages
