+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.2"
     defaultConfig {
-        applicationId "com.alexzh.jetpackcomposeworkshop"
+        applicationId "com.alexzh.coffeedrinks"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
