+++ app/src/main/java/com/alexzh/coffeedrinks/data/CoffeeDrink.kt
@@ -1,0 +1,15 @@
+package com.alexzh.coffeedrinks.data
+
+import androidx.annotation.DrawableRes
+
+data class CoffeeDrink(
+    val id: Long,
+    val name: String,
+    @DrawableRes val imageUrl: Int,
+    @DrawableRes val backgroundImage: Int,
+    val description: String,
+    val ingredients: String,
+    val orderDescription: String,
+    val price: Double,
+    var isFavourite: Boolean
+)
