+++ app/src/main/java/com/alexzh/coffeedrinks/data/CoffeeDrinkRepository.kt
@@ -1,0 +1,12 @@
+package com.alexzh.coffeedrinks.data
+
+import com.alexzh.coffeedrinks.data.CoffeeDrink
+
+interface CoffeeDrinkRepository {
+
+    fun getCoffeeDrinks(): List<CoffeeDrink>
+
+    fun getCoffeeDrink(id: Long): CoffeeDrink?
+
+    fun updateCoffeeDrink(coffeeDrink: CoffeeDrink): Boolean
+}
