--- app/src/main/java/com/alexzh/jetpackcomposeworkshop/data/CoffeeDrinkRepository.kt
@@ -1,10 +1,0 @@
-package com.alexzh.jetpackcomposeworkshop.data
-
-interface CoffeeDrinkRepository {
-
-    fun getCoffeeDrinks(): List<CoffeeDrink>
-
-    fun getCoffeeDrink(id: Long): CoffeeDrink?
-
-    fun updateCoffeeDrink(coffeeDrink: CoffeeDrink): Boolean
-}
