--- app/src/main/java/com/alexzh/jetpackcomposeworkshop/ui/coffeedrinks/model/CoffeeDrinksModel.kt
@@ -1,9 +1,0 @@
-package com.alexzh.jetpackcomposeworkshop.ui.coffeedrinks.model
-
-import androidx.compose.Model
-import androidx.compose.frames.ModelList
-
-@Model
-object CoffeeDrinksModel {
-    val coffeeDrinks = ModelList<CoffeeDrinkItem>()
-}
