+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">JetpackComposeWorkshop</string>
+    <string name="app_name">Coffee drinks</string>
 </resources>
